+++ app/src/main/java/org/simple/clinic/main/TheActivityComponent.kt
@@ -44,11 +44,11 @@ import org.simple.clinic.navigation.di.FragmentScreenKeyModule
 import org.simple.clinic.navigation.v2.Router
 import org.simple.clinic.newentry.PatientEntryScreen
 import org.simple.clinic.newentry.country.di.InputFieldsFactoryModule
-import org.simple.clinic.onboarding.OnboardingScreenInjector
 import org.simple.clinic.recentpatient.RecentPatientsScreen
 import org.simple.clinic.recentpatientsview.RecentPatientsView
 import org.simple.clinic.removeoverdueappointment.RemoveOverdueAppointmentScreen
 import org.simple.clinic.navigation.v2.ScreenResultBus
+import org.simple.clinic.onboarding.OnboardingScreen
 import org.simple.clinic.scanid.ScanSimpleIdScreen
 import org.simple.clinic.scanid.scannedqrcode.ScannedQrCodeSheet
 import org.simple.clinic.scheduleappointment.ScheduleAppointmentSheet
@@ -79,7 +79,7 @@ import org.simple.clinic.widgets.PatientSearchResultItemView
 
 @Subcomponent(modules = [TheActivityModule::class])
 interface TheActivityComponent :
-    OnboardingScreenInjector,
+    OnboardingScreen.Injector,
     MedicalHistorySummaryViewInjector,
     DrugSummaryViewInjector,
     BloodSugarSummaryViewInjector,